Refactor: Class 구조적용 및 코드정리 (import-export)
