+++ README/STEP3.md
@@ -40,9 +40,9 @@ TAB 을 이용한 UI는 화면에 더 많은 데이터를 노출 할 수 있는 
 - [x] JS코드로 객체들 만들고 추가하기
 - [x] AJAX 데이터 넣기
 - [x] 본 JS코드에 적용
-- [ ] 클릭이벤트와 연동
-  - [ ] 클릭한 엘리먼트 배경색 및 색깔 바꾸기
-- [ ] 맨 처음에 랜덤으로 보여주기
+- [x] 클릭이벤트와 연동
+  - [x] 클릭한 엘리먼트 배경색 및 색깔 바꾸기
+- [x] 맨 처음에 랜덤으로 보여주기
 
 <br/><br/>
 
@@ -152,7 +152,14 @@ TAB 을 이용한 UI는 화면에 더 많은 데이터를 노출 할 수 있는 
 
 8. **JavaScript 에서 click 이벤트를 적용 후, CSS hover 동작안함 **
 
-   `!important` 로 해결했으나, 다른 방법을 찾고싶음
+   `!important` 로 해결했으나, 다른 방법을 **찾고싶음**
+
+9. **Class 구조 적용 (ES6)**
+
+   export { 'class name' }
+
+   import { 'class name' } from '파일위치'
+
 
  <br/><br/>
