--- create-element-test.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-    <link rel="stylesheet" href="lib/reset.css">
-    <link rel="stylesheet" href="css/main.css">
-</head>
-<body>
-    <h1>노드들을 잘 만들고 AJAX로 가져온 데이터를 잘 만드는지 테스트합니다</h1>
-</body>
-
-<!-- <script src="./js/test-ajax.js"></script> -->
-<!-- <script src="./js/best-dish.js"></script> -->
-<script src="./js/ajax-test.js"></script>
-
-</html>
