+++ main.html
@@ -527,7 +527,6 @@
     </div>
   </body>
 
-  <script src="./js/ajax-test.js"></script>
-  <script src="main-test.js"></script>
+  <script type="module" src="http://localhost:5500/app.js"></script>
 
 </html>
